requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.4.0
tqdm>=4.64.0
PyYAML>=6.0
scipy>=1.10.0
heapq-max>=0.21
jupyter>=1.0.0
ipykernel>=6.20.0
